Se quita comentarios del codigo y se modulariza
