+++ src/constants/constantes.js
@@ -1,0 +1,27 @@
+export const colorMap = {
+  var: '#F92672',
+  while: '#F92672',
+  if: '#F92672',
+  else: '#F92672',
+  0: '#AD80FD',
+  true: '#AD80FD',
+  '¡Feliz Cumple': '#D4CA6C',
+};
+
+export const estilosTexto = {
+  pseudoCodigo: {
+    font: '30px monospace',
+    fillStyle: 'white',
+    textAlign: 'left',
+  },
+  textoDiaCumple: {
+    font: 'bold 120px sans-serif',
+    fillStyle: '#000000',
+    textAlign: 'center',
+  },
+  textoMesCumple: {
+    font: 'bold 40px sans-serif',
+    fillStyle: '#808080',
+    textAlign: 'left',
+  },
+};
