--- src/pages/Principal.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import '../styles/Principal.css';
-
-const Principal = () => {
-  return (
-    <div>Principal</div>
-  )
-}
-
-export default Principal
