+++ src/utils/manejadorFechas.js
@@ -1,0 +1,7 @@
+export const getMesNombre = (mes) => {
+  const meses = [
+    "Enero", "Febrero", "Marzo", "Abril", "Mayo", "Junio",
+    "Julio", "Agosto", "Septiembre", "Octubre", "Noviembre", "Diciembre"
+  ];
+  return meses[mes - 1];
+};
